abandoning approach; committing just in case i want to try it again

+++ core/src/main/java/ubc/team09/eval/HeuristicMethod.java
@@ -4,11 +4,11 @@ import ubc.team09.state.State;
 
 public interface HeuristicMethod {
 	/**
-	 * Returns a number in the range of [-1, +1] where lower values indicate
+	 * Returns a number in the range of (-∞, +∞) where lower values indicate
 	 * a more favorable position for Black, and greater values indicate a more
 	 * favorable position for White.
 	 * 
 	 * @param state The board state to evaluate.
 	 */
-	public double evaluate(State state);
+	public int evaluate(State state);
 }
